Completion pass 2: rebuild the 3R detector against its measured evidence, close the fourteen post-completion findings
The detector-baseline measurement committed at 8d72b8c was negative on
every axis; this pass resolves what it showed, plus every other issue the
post-completion audit left open — none by documentation alone.

* 3R timing detector rebuilt: point alarms score against the PRIOR window
  (the old post-update z was provably capped below 3.0, making four of six
  sigma profiles and 'critical' severity unreachable) at
  max(threshold_sigma, empirically calibrated (1-alarm_budget) quantile);
  sustained shifts raise edge-triggered sign-CUSUM events against a
  reference locked at 200 samples, re-baselining after persistence;
  severity is capped at 'warning' until calibrated; profiles now cover the
  operation names production actually emits and TimingAnomaly gains a
  `kind` field. Measured on real Ed25519 wall-clock timings: clean
  point-alarm rate 0.78-1.0% against the 1% budget (was 12.52%), 30% shift
  alerted in ~11-20 samples with 0.93+ regime coverage (was 17.6% recall),
  alarms strictly monotone in both budget and sigma. The 1,807-line stale
  detector copy under tools/monitoring/ is deleted.
* benchmarks/detector_baseline_eval.py drives the real monitor (the old
  harness re-implemented a stronger rule than production shipped), derives
  its tie band from cross-seed spread instead of _TIE_F1=0.02, labels its
  ground truth as synthetic injection over measured timings, and runs in
  CI as a gate (ci.yml ubuntu/3.11) inside the mypy --strict scope. The
  honest baseline result is stated, not spun: at a matched budget a top-N
  KNN oracle ranks isolated spikes as well or slightly better (mean F1
  ratio 0.80-0.98); the machinery's earned value is the calibrated budget
  and the shift path.
* dudect: ChaCha20-Poly1305 and AES-GCM each gain a forgery-position lane
  (two forgeries over a 64-byte payload through the corrected masked
  control flow, differing only in the mismatch byte); both stale
  "if (ct_len > 0)" lane comments corrected to the bounded_len masking the
  implementations actually ship; Argon2/agent-binding position coverage
  reasoned in-tree (straight-line fail paths; the shared memcmp lane
  carries the property). 27 lanes, all green at 100K measurements.
* Kyber -Wconversion driven to zero and frozen: barrett_reduce rewritten
  with int32 intermediates and one exhaustively-proven cast (65,536/65,536
  bit-identical) in the generic, AVX2, NEON (latent pre-fix form), SVE2
  and test-reference variants; clang's 42 implicit-int-conversion findings
  in the compress/decompress packers fixed in the d=10 branch's own
  explicit-cast style; the strict-warnings job now fails on any non-vendor
  warning beyond the two documented extension classes, with pipefail.
* CodeQL alert 620 resolved at source with the established _explode()
  pattern; no dismissal, no suppression.
* Lines-of-code figures gated on BOTH columns plus the Scope Composition
  table and every prose restatement, measured from git ls-files with
  newline-byte counts; `python tools/update_docs.py --loc` regenerates
  every gated figure from the same functions the checker verifies with.
  Makefiles join the whole-project scope (the two published commands
  disagreed by 502 lines); the find-based commands are retired.
* Windows runs the real SoftHSM2 token lifecycle: the "no maintained
  Windows package" claim was checked and is false (choco softhsm.install
  2.5.0 / Disig MSI); both workflows provision it fail-loud, the [hsm]
  extra installs on Windows, the probe and PKCS11_PATHS know the DLL
  location, and the win32 exemption in the provisioning guard is removed.
* Linux wheels are re-signed AFTER auditwheel repair
  (tools/resign_wheel.py chained in CIBW_REPAIR_WHEEL_COMMAND_LINUX) so
  the integrity artefact binds shipped bytes by construction; the
  .text-only digest alternative is rejected on the record. SECURITY.md's
  two stale binding paragraphs corrected (delocate DID rewrite Mach-O
  load commands; the pre-import gate exists).
* tools/verify_install_oob.py: standalone stdlib-only out-of-band
  verifier (hand-written FIPS 202 SHA3-256, FIPS 180-4 SHA-512, RFC 8032
  Ed25519 verify, self-KAT'd fail-closed) that verifies an installed
  tree's signature, source digest, native/binding digests and every
  cached .pyc from OUTSIDE the checker-poisoning boundary — the one
  genuine engineering improvement the recorded reassessment left open.
* README canonical-host tables: the falsified "re-measured at release
  time" sentence is withdrawn; every row's 4.x provenance is stated and
  the missing AVX-512/VAES/VPCLMULQDQ 5.0.0 measurement is an explicit
  release-time action, not an implied fact.
* Repository noise removed at root cause: the slow-unit-* libFuzzer
  artifact committed in da63d5c deleted, artifact family gitignored.

Verified on this tree: C suites 63/63 (gcc and clang, strict flags, both
warning gates green), dudect 27/27 at 100K, full Python suite green,
mypy --strict clean over 221 files, black/ruff clean, every tools/check_*
gate green, detector efficacy gates green on real timings, valgrind
memcheck / MemorySanitizer / ThreadSanitizer executed locally against
this exact tree, and the OOB verifier passes end-to-end with tamper
cases refused.

+++ src/c/sve2/ama_kyber_sve2.c
@@ -95,9 +95,9 @@ static inline int16_t barrett_reduce_scalar(int16_t a) {
      * comment above claimed it "matches generic C reference exactly"; it did
      * not, and nothing noticed because AMA_ENABLE_SVE2 was off in every CI
      * configuration, so this code had never run under the test that pins it. */
-    int16_t t = (int16_t)(((int32_t)KYBER_BARRETT_V * a) >> 26);
+    int32_t t = ((int32_t)KYBER_BARRETT_V * (int32_t)a) >> 26;
     t *= KYBER_Q;
-    return a - t;
+    return (int16_t)(a - t);
 }
 
 /* ============================================================================
